fix: log playwright browser download progress on stderr
If a browser is downloaded during a test run, playwright will log
on stdout.

This can interfere with the calling code if it's trying to parse
the output of the program as, for example, JSON.

The fix is to temporarily replace `console.log` with `console.error`
as that's what the [logPolitely](https://github.com/microsoft/playwright/blob/718bd9b35fd206245401a9ecb320289f427592d9/packages/playwright-core/src/server/registry/browserFetcher.ts#L120)
function uses to give feedback to the user.

The intention of the playwright maintainers is [not for this
functionality to be used during a test run](https://github.com/microsoft/playwright/issues/32487#issuecomment-2337666079)
so it's unlikely to move to stderr otherwise.